requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.0
requests==2.26.0
Django==3.2
django-filter==23.2
django-extensions==3.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2